Move D.Client.Compat.Environment into Cabal
